requirements.txt
Django==1.9.5
django-taggit==0.18.1
